Experimental: endless paging and pjax page loading
This is a significant piece of work that allows endless paging of question pages
and PJAX page loading. Currently, the admin pages and a few other limited areas
are the only pieces of the application that fall back to standard requests.

Both features are enabled via the 'web' settings panel. This also includes a
fixed top navigation bar that sticks to the top of the page while scrolling.
JavaScript has yet to be added to ensure that the navigation buttons match the
current page state, but otherwise this looks pretty solid.